Reducing verbosity

+++ tasks/aws.py
@@ -28,12 +28,10 @@ def build_lambdas(ctx):
 
     build_dir = join(PROJ_ROOT, "lambda_build")
 
-    rmtree(build_dir)
-
     _build_cmake_project(build_dir, [
         "-DFAASM_BUILD_TYPE=lambda",
         "-DCMAKE_BUILD_TYPE=Debug",
-    ])
+    ], clean=True)
 
 
 @task
@@ -120,20 +118,22 @@ def build_lambda_runtime(ctx):
         mkdir(INSTALL_PATH)
 
     # AWS C++ runtime
-    # NOTE: building in debug mode enables verbose logging
+    # NOTE: building in debug mode enables verbose logging by default
+    # Verbosity can be set on release builds (1 = info, 2 = debug, 3 = all)
     _set_up_cmake_project(
         "https://github.com/awslabs/aws-lambda-cpp-runtime.git",
         "aws-lambda-cpp-runtime",
         [
-            "-DCMAKE_BUILD_TYPE=Debug",
-            "-DLOG_VERBOSITY=3",
+            "-DCMAKE_BUILD_TYPE=Release",
+            "-DLOG_VERBOSITY=1",
             "-DBUILD_SHARED_LIBS=OFF",
         ],
-        RUNTIME_VERSION
+        RUNTIME_VERSION,
+        clean=True
     )
 
 
-def _set_up_cmake_project(repo_url, dir_name, cmake_args, version):
+def _set_up_cmake_project(repo_url, dir_name, cmake_args, version, clean=False):
     checkout_path = join(FAASM_HOME, dir_name)
 
     if not exists(checkout_path):
@@ -146,17 +146,17 @@ def _set_up_cmake_project(repo_url, dir_name, cmake_args, version):
         call("git checkout {}".format(version), cwd=checkout_path, shell=True)
 
     build_dir = join(checkout_path, "build")
-    if exists(build_dir):
-        rmtree(build_dir)
-
-    mkdir(build_dir)
-
-    _build_cmake_project(build_dir, cmake_args)
+    _build_cmake_project(build_dir, cmake_args, clean=clean)
 
     call("make install", cwd=build_dir, shell=True)
 
 
-def _build_cmake_project(build_dir, cmake_args):
+def _build_cmake_project(build_dir, cmake_args, clean=False):
+    # Nuke if required
+    if clean and exists(build_dir):
+        rmtree(build_dir)
+
+    # Create build dir if necessary
     if not exists(build_dir):
         mkdir(build_dir)
